backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
PyJWT>=2.6.0
requests>=2.28.0
python-decouple>=3.8

# Database
psycopg2-binary==2.9.7  # For PostgreSQL (optional)


# JWT Authentication
PyJWT==2.8.0

# AI/ML for future features
openai==0.28.1
langchain==0.0.335

# Utilities
python-decouple==3.8  # For environment variables
python-dateutil==2.8.2
requests==2.31.0

# Development
django-debug-toolbar==4.2.0  # For development debugging